fixed all of that, looks cool
