+++ Core/DI/MIN.Core.DI/FeatureCollection/MinFeatureCollection.cs
@@ -1,0 +1,40 @@
+﻿using MIN.Core.Events.Contracts;
+using MIN.Core.Services.Contracts.Interfaces.Rooms;
+using MIN.Core.Stores.Contracts.Interfaces;
+
+namespace MIN.Core.DI.FeatureCollection;
+
+/// <inheritdoc cref="ICoreFeatureCollection"/>
+public class CoreFeatureCollection : ICoreFeatureCollection
+{
+    /// <inheritdoc cref="IRoomConnector"/>
+    public IRoomConnector RoomConnector { get; }
+
+    /// <inheritdoc cref="IRoomHoster"/>
+    public IRoomHoster RoomHoster { get; }
+
+    /// <inheritdoc cref="IRoomStore"/>
+    public IRoomStore RoomStore { get; }
+
+    /// <inheritdoc cref="IRoomFactory"/>
+    public IRoomFactory RoomFactory { get; }
+
+    /// <inheritdoc cref="IEventBus"/>
+    public IEventBus EventBus { get; }
+
+    /// <summary>
+    /// Инициализирует новый экземпляр <see cref="CoreFeatureCollection"/>
+    /// </summary>
+    public CoreFeatureCollection(IRoomConnector roomConnector,
+        IRoomHoster roomHoster,
+        IRoomStore roomStore,
+        IRoomFactory roomFactory,
+        IEventBus eventBus)
+    {
+        RoomConnector = roomConnector;
+        RoomHoster = roomHoster;
+        RoomStore = roomStore;
+        RoomFactory = roomFactory;
+        EventBus = eventBus;
+    }
+}
