+++ DI/MIN.DI/FeatureCollection/MinFeatureCollection.cs
@@ -1,0 +1,41 @@
+﻿using MIN.Chat.DI.FeatureCollection;
+using MIN.Core.DI.FeatureCollection;
+using MIN.Discovery.DI.FeatureCollection;
+using MIN.Helpers.DI.FeatureCollection;
+
+namespace MIN.DI.FeatureCollection;
+
+/// <inheritdoc cref="IMinFeatureCollection"/>
+public class MinFeatureCollection : IMinFeatureCollection
+{
+    /// <inheritdoc />
+    public IChatFeatureCollection Chat { get; }
+
+    /// <inheritdoc />
+    public ICoreFeatureCollection Core { get; }
+
+    /// <inheritdoc />
+    public IDiscoveryFeatureCollection Discovery { get; }
+
+    /// <inheritdoc />
+    public IHelperFeatureCollection Helper { get; }
+
+    /// <inheritdoc />
+    public Version Version { get; }
+
+    /// <summary>
+    /// Инициализирует новый экземпляр <see cref="MinFeatureCollection"/>
+    /// </summary>
+    public MinFeatureCollection(IChatFeatureCollection chat,
+        ICoreFeatureCollection core,
+        IDiscoveryFeatureCollection discovery,
+        IHelperFeatureCollection helper,
+        Version version)
+    {
+        Chat = chat;
+        Core = core;
+        Discovery = discovery;
+        Helper = helper;
+        Version = version;
+    }
+}
