+++ DI/MIN.DI/MinModule.cs
@@ -1,8 +1,9 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using System.Reflection;
+using Microsoft.Extensions.DependencyInjection;
 using MIN.Chat.DI;
-using MIN.Common.Mvc;
 using MIN.Common.Mvc.Extensions;
 using MIN.Core.DI;
+using MIN.DI.FeatureCollection;
 using MIN.Discovery.DI;
 using MIN.Helpers.DI;
 
@@ -11,14 +12,19 @@ namespace MIN.DI;
 /// <summary>
 /// Модуль регистрации зависимостей для MIN
 /// </summary>
-public class MinModule : Module
+public class MinModule : Common.Mvc.Module
 {
     /// <inheritdoc />
     protected override void Load(IServiceCollection services)
     {
+        var appVersion = Assembly.GetEntryAssembly()?.GetName().Version ?? new Version(0, 0, 0, 0);
+        services.AddSingleton(appVersion);
+
         services.RegisterModule<HelpersModule>();
         services.RegisterModule<CoreModule>();
         services.RegisterModule<ChatModule>();
         services.RegisterModule<DiscoveryModule>();
+
+        services.RegisterAsImplementedInterfaces<MinFeatureCollection>(ServiceLifetime.Singleton);
     }
 }
