+++ Desktop/MIN.Desktop/Infrastructure/Services/NavigationService.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using MIN.Desktop.Contracts.Interfaces;
+using MIN.Desktop.Contracts.Views.Forms;
 using MIN.Desktop.Contracts.Views.PanelViews;
 using MIN.Desktop.Contracts.Views.PanelViews.Models;
 
@@ -8,17 +9,24 @@ namespace MIN.Desktop.Infrastructure.Services;
 /// <inheritdoc cref="INavigationService"/>
 public class NavigationService : INavigationService
 {
-    private readonly Panel mainPanel;
-    private readonly Panel sidePanel;
+    BaseForm INavigationService.Parent { get; set; } = null!;
+
+    /// <inheritdoc />
+    public SplitContainer SplitContainer { get; set; } = null!;
+
+    /// <inheritdoc />
+    public Panel MainPanel { get; set; } = null!;
+
+    /// <inheritdoc />
+    public Panel SidePanel { get; set; } = null!;
+
     private readonly IServiceProvider provider;
 
     /// <summary>
     /// Инициализирует новый экземпляр <see cref="NavigationService"/>
     /// </summary>
-    public NavigationService(Panel mainPanel, Panel sidePanel, IServiceProvider provider)
+    public NavigationService(IServiceProvider provider)
     {
-        this.mainPanel = mainPanel;
-        this.sidePanel = sidePanel;
         this.provider = provider;
     }
 
@@ -45,15 +53,16 @@ public class NavigationService : INavigationService
         switch (panel.PanelType)
         {
             case Contracts.Enums.PanelType.Main:
-                mainPanel.Controls.Clear();
-                mainPanel.Controls.Add(panel);
+                MainPanel.Controls.Clear();
+                SplitContainer.Panel2MinSize = panel.MinimumSize.Width;
+                MainPanel.Controls.Add(panel);
                 break;
 
             case Contracts.Enums.PanelType.Side:
-                sidePanel.Controls.Clear();
-                sidePanel.Controls.Add(panel);
+                SidePanel.Controls.Clear();
+                SplitContainer.Panel1MinSize = panel.MinimumSize.Width;
+                SidePanel.Controls.Add(panel);
                 break;
-
         }
 
         panel.Dock = DockStyle.Fill;
