+++ Helpers/MIN.Helpers/Services/IdentityService.cs
@@ -1,4 +1,5 @@
 ﻿using MIN.Core.Entities.Contracts.Interfaces;
+using MIN.Core.Entities.Contracts.Models;
 using MIN.Helpers.Contracts.Interfaces;
 
 namespace MIN.Helpers.Services;
@@ -6,7 +7,10 @@ namespace MIN.Helpers.Services;
 /// <inheritdoc />
 public sealed class IdentityService : IIdentityService
 {
-    private IParticipantData currentParticipant = null!;
+    private IParticipantData currentParticipant = new ParticipantInfo()
+    {
+        Name = "Ты"
+    };
 
     IParticipantData IIdentityService.SelfPartcipant => currentParticipant;
