+++ Infrastructure/Chat/MIN.Chat.DI/FeatureCollection/ChatFeatureCollection.cs
@@ -1,0 +1,18 @@
+﻿using MIN.Chat.Services.Contracts.Interfaces;
+
+namespace MIN.Chat.DI.FeatureCollection;
+
+/// <inheritdoc cref="IChatFeatureCollection"/>
+public class ChatFeatureCollection : IChatFeatureCollection
+{
+    /// <inheritdoc />
+    public IChatService ChatService { get; }
+
+    /// <summary>
+    /// Инициализирует новый экземпляр <see cref="ChatFeatureCollection"/>
+    /// </summary>
+    public ChatFeatureCollection(IChatService chatService)
+    {
+        ChatService = chatService;
+    }
+}
